backend/atomic/student/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-restx==1.3.0
requests==2.32.3
supabase==2.15.1
python-dotenv==1.1.0
PyJWT==2.10.1
cryptography==44.0.2
